refactor: split ObjectInitializerTests into partial classes
